reorg: pull bsr into separate source files
