+++ CMakeLists_files.cmake
@@ -157,7 +157,8 @@ list (APPEND MAIN_SOURCE_FILES
   opm/simulators/linalg/FlexibleSolver5.cpp
   opm/simulators/linalg/FlexibleSolver6.cpp
   opm/simulators/linalg/FlowLinearSolverParameters.cpp
-  opm/simulators/linalg/HaugenLabs.c
+  opm/simulators/linalg/mixed/HaugenLabs.c
+  opm/simulators/linalg/mixed/bsr.c
   opm/simulators/linalg/ISTLSolver.cpp
   opm/simulators/linalg/MILU.cpp
   opm/simulators/linalg/ParallelIstlInformation.cpp
